+++ KITforg_labs/Assets/Plugins/KITforge Labs/KF_HierarchyKit/Editor/KF_HierarchyKitInitializer.cs
@@ -8,6 +8,7 @@ namespace KITforgeLabs.Editor.HierarchyKit
         static KF_HierarchyKitInitializer()
         {
             EditorApplication.hierarchyWindowItemOnGUI += KF_HierarchyKitRenderer.OnHierarchyItemGUI;
+            EditorApplication.hierarchyChanged += KF_HierarchyKitRenderer.ClearNameCache;
         }
     }
 }